Add seperate button for cancel offer action and define modals inside buttons
